LayoutTests/fast/css/getComputedStyle/getComputedStyle-outline-shorthand-expected.txt

Test to make sure outline shorthand property returns CSSValueList properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS computedStyle.getPropertyValue('outline') is '5px solid rgb(255, 0, 0)'
PASS computedStyle.getPropertyCSSValue('outline').toString() is '[object CSSValueList]'
PASS computedStyle.getPropertyCSSValue('outline').length is 3
PASS computedStyle.getPropertyCSSValue('outline').cssText is '5px solid rgb(255, 0, 0)'
PASS computedStyle.getPropertyCSSValue('outline').item(0).getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 5
PASS computedStyle.getPropertyCSSValue('outline').item(1).getStringValue() is 'solid'
PASS computedStyle.getPropertyCSSValue('outline').item(2).getRGBColorValue().red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 255
PASS computedStyle.getPropertyCSSValue('outline').item(2).getRGBColorValue().green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS computedStyle.getPropertyCSSValue('outline').item(2).getRGBColorValue().blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS computedStyle.getPropertyValue('outline') is '32px solid rgb(0, 0, 0)'
PASS computedStyle.getPropertyCSSValue('outline').toString() is '[object CSSValueList]'
PASS computedStyle.getPropertyCSSValue('outline').length is 3
PASS computedStyle.getPropertyCSSValue('outline').cssText is '32px solid rgb(0, 0, 0)'
PASS computedStyle.getPropertyCSSValue('outline').item(0).getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 32
PASS computedStyle.getPropertyCSSValue('outline').item(1).getStringValue() is 'solid'
PASS computedStyle.getPropertyCSSValue('outline').item(2).getRGBColorValue().red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS computedStyle.getPropertyCSSValue('outline').item(2).getRGBColorValue().green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS computedStyle.getPropertyCSSValue('outline').item(2).getRGBColorValue().blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS computedStyle.getPropertyValue('outline') is '4px none rgb(255, 0, 0)'
PASS computedStyle.getPropertyCSSValue('outline').toString() is '[object CSSValueList]'
PASS computedStyle.getPropertyCSSValue('outline').length is 3
PASS computedStyle.getPropertyCSSValue('outline').cssText is '4px none rgb(255, 0, 0)'
PASS computedStyle.getPropertyCSSValue('outline').item(0).getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 4
PASS computedStyle.getPropertyCSSValue('outline').item(1).getStringValue() is 'none'
PASS computedStyle.getPropertyCSSValue('outline').item(2).getRGBColorValue().red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 255
PASS computedStyle.getPropertyCSSValue('outline').item(2).getRGBColorValue().green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS computedStyle.getPropertyCSSValue('outline').item(2).getRGBColorValue().blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS successfullyParsed is true

TEST COMPLETE
